discord_bot/backend/app/requirements.txt
aiofiles==24.1.0
alembic==1.12.1
annotated-types==0.7.0
bcrypt==4.2.1
fastapi==0.104.1
httpx==0.27.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.5.2
# pydantic-core pin removed: let pydantic pull its compatible version.
# The old pin (2.14.5) had no wheel for Python 3.12+ and failed to build on Render.
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
shortuuid==1.0.11
SQLAlchemy==2.0.23
supabase==2.10.0
uvicorn==0.24.0
